Added getname as utils and remove setname from precall
