+++ template/src/components/precall/usePreCall.tsx
@@ -14,16 +14,12 @@ import {createHook} from 'fpe-implementation';
 import {ApolloError} from '@apollo/client';
 
 export interface PreCallContextInterface {
-  username: string;
-  setUsername: React.Dispatch<React.SetStateAction<string>>;
   callActive: boolean;
   setCallActive: React.Dispatch<React.SetStateAction<boolean>>;
   error?: ApolloError;
 }
 
 const PreCallContext = createContext<PreCallContextInterface>({
-  username: '',
-  setUsername: () => {},
   callActive: false,
   setCallActive: () => {},
 });
